ref: move init to single file, update css
