+++ src/instruments/src/DCDU/components/DcduContainer.jsx
@@ -1,16 +1,16 @@
 /* eslint-disable react/prop-types */
-import './DcduContainer.style.css';
+import './DcduContainer.css';
 
 export function DcduContainer({
   top, left, middle, right,
 }) {
   return (
     <>
-      <div className="dcdu__container">
-        <div className="dcdu__container--top">{top}</div>
-        <div className="dcdu__container--left">{left}</div>
-        <div className="dcdu__container--middle">{middle}</div>
-        <div className="dcdu__container--right">{right}</div>
+      <div className="dcdu-container">
+        <div className="dcdu-container-top">{top}</div>
+        <div className="dcdu-container-left">{left}</div>
+        <div className="dcdu-container-middle">{middle}</div>
+        <div className="dcdu-container-right">{right}</div>
       </div>
     </>
   );
