+++ src/instruments/src/DCDU/pages/Idle/Idle.jsx
@@ -1,10 +1,10 @@
-import './Idle.style.css';
+import './Idle.css';
 
 import { DcduContainer } from '../../components/DcduContainer';
 
 export function Idle() {
   const top = (
-    <div className="dcdu__idle">
+    <div className="dcdu-idle">
       NO ACTIVE ATC
     </div>
   );
