--- src/instruments/src/DCDU/pages/SelfTest/SelfTest.style.css
@@ -1,18 +1,0 @@
-@import url("/CSS/A32NX_Display_Common.css");
-
-.dcdu__selfTest--container {
-  width: 100%;
-  min-height: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-family: ECAMFontRegular;
-  color: var(--displayGreen);
-  font-size: 15pt;
-  background: radial-gradient(
-    ellipse at center,
-    rgba(13, 20, 35, 1) 0%,
-    rgb(13, 20, 35) 100%
-  );
-}
